api/requirements.txt
Flask==2.3.3
flask-cors==4.0.1
python-decouple==3.8
requests==2.26.0
beautifulsoup4==4.9.3

